feat: Phase 5 - Presentation層リファクタリング (Tasks 6.1-6.3 部分完了)
**Task 6.1: cli.ts の分割 ✅**
- src/cli.ts を10行のシンプルなラッパーに削減 (996行 → 10行)
- 実際のロジックを src/presentation/cli.ts に移行 (約900行)
- バージョン取得とconfig関連ユーティリティを分離
  - src/presentation/cli/version.ts
  - src/presentation/cli/config.ts
- グローバルエラーハンドリング実装
- すべてのコマンド登録を関数化して整理

**Task 6.2: init コマンドの分割と移行 ✅**
- src/commands/init.ts を10行のシンプルなラッパーに削減 (695行 → 10行)
- 5つのモジュールに分割:
  - validation.ts (80行) - バリデーション関数
  - prompts.ts (180行) - 対話型プロンプト
  - templates.ts (200行) - テンプレート処理
  - setup.ts (230行) - セットアップ実行ロジック
  - handler.ts (90行) - メインハンドラー
- 各モジュールの責務を明確化
- 再利用性と保守性を向上

**Task 6.3: 他のコマンドハンドラーの移行 (部分完了) ⏳**
- spec-archive コマンドを src/presentation/commands/spec/archive.ts に移行
- spec-list コマンドを src/presentation/commands/spec/list.ts に移行
- 元のコマンドファイルをシンプルなラッパーに変更

**テスト結果**
- ✅ 925 tests passed, 5 skipped
- ✅ ビルド成功
- ✅ すべての機能が正常に動作

**次のステップ**
- Task 6.3 の継続: jira-sync, confluence-sync, workflow コマンドの移行
- Task 6.4: Multi-Repo コマンドの移行
- Task 6.5: フォーマッタの作成
- Task 6.6: 対話型UIの移行
- Task 6.7: Presentation層のテストと統合確認
